salvataggio

+++ libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: fgolino <fgolino@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/18 16:36:33 by fgolino           #+#    #+#             */
-/*   Updated: 2023/01/19 18:27:07 by fgolino          ###   ########.fr       */
+/*   Updated: 2023/01/23 18:04:59 by fgolino          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(void *s1, void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
