+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: fgolino <fgolino@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 17:12:55 by fgolino           #+#    #+#             */
-/*   Updated: 2023/01/19 14:59:36 by fgolino          ###   ########.fr       */
+/*   Updated: 2023/01/23 18:13:05 by fgolino          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(char *str1, char *str2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
